+++ packages/featureserver/test/query.js
@@ -21,6 +21,7 @@ describe('Query operatons', function () {
   describe('when using an outSR', function () {
     it('should translate the data properly', function () {
       const response = FeatureServer.query(data, { outSR: { latestWkid: 3857 }, limit: 1, returnGeometry: true })
+      response.geometryType.should.equal('esriGeometryPoint')
       response.features.length.should.equal(1)
       response.features[0].geometry.x.should.equal(-11682713.391976157)
       response.features[0].geometry.y.should.equal(4857924.005275469)