+++ packages/featureserver/CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [1.2.3] - 2017-02-27
 ### Fixed
 * Correct response on /FeatureServer endpoint
 
@@ -70,6 +70,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [1.0.0] - 2017-01-25
 * Initial Release
 
+[1.2.3]: https://github.com/featureserver/featureserver/compare/v1.2.3...v1.2.2
 [1.2.2]: https://github.com/featureserver/featureserver/compare/v1.2.1...v1.2.2
 [1.2.1]: https://github.com/featureserver/featureserver/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/featureserver/featureserver/compare/v1.1.1...v1.2.0